> moved messageHandler as per #23
