+++ handlers/message/messageHandler.js
@@ -1,7 +1,7 @@
-const CONFIG = require("../modules/config");
-const commandHandler = require("./command/commandHandler").handler;
-const handleDataCheck = require("../checks/handleData").check;
-const dbInt = require("../db/interface");
+const CONFIG = require("../../modules/config");
+const commandHandler = require("../command/commandHandler").handler;
+const handleDataCheck = require("../../checks/handleData").check;
+const dbInt = require("../../db/interface");
 
 module.exports = {
     handler: async function (handleData) {
